// src/types/api_res/RecentList/RecentListGet.ts
export interface RecentListGet {
    success: boolean;
    message: string;
    errorCode: number;
    data: Data[];
}

export interface Data {
    product_today_seq: number;
    product_cd: string;
    category1_cd: null;
    category2_cd: null;
    category3_cd: null;
    user_id: string;
    session_id: string;
    use_yn: string;
    cust_seq: number;
    reg_date: string;
    product_code: null;
    product_type: string;
    product_state: string;
    brand_cd: string;
    group_yn: string;
    group_cd: string;
    product_nm: string;
    product_etc_nm: string;
    product_deal_nm: string;
    product_nm_eng: null;
    product_nm_mb: null;
    model_name: null;
    product_stock: null;
    coupon_use_yn: string;
    reserve_use_yn: string;
    reserve_give_yn: string;
    new_yn: string;
    supply_price: number;
    sale_price: number;
    fee_rate: number;
    reserve_rate: number;
    keywd: string;
    product_content: null;
    icon: string;
    gid: null;
    main_img: null;
    sub_img1: null;
    sub_img2: null;
    sub_img3: null;
    sub_img4: null;
    free_trans_yn: number;
    discount_gb: string;
    product_national_cd: null;
    mov_code: null;
    product_relation_yn: string;
    best_product_display: string;
    best_review_display: string;
    product_url: null;
    write_use_yn: string;
    write_title: string;
    order_limit_cnt: number;
    order_mini_quantiry: string;
    shipping_yn: string;
    shipping_gudie: string;
    return_yn: string;
    return_guide: string;
    exchange_yn: string;
    exchange_guide: string;
    od: number;
    vw_cnt: number;
    hits: number;
    review_cnt: number;
    order_cnt: number;
    del_yn: string;
    mod_date: string;
    mod_id: string;
    point: number;
    review_pt_yn: string;
    hidden_yn: string;
    main_title: null;
    sub_title: null;
    product_main_img: string;
    wish_click_on: string;
}